refactor(nemotron-multilingual): split +Pipeline.swift by concern
The 1,482-line +Pipeline.swift is split into three files of the same
extension (no logic change, no API change):
  +Pipeline.swift  — processChunk orchestration (prep/encoder/decode/pipelining)
  +Decode.swift    — runSpeculativeBlankDecodeV2, argmaxFrame, tokenizerPiece
  +Buffers.swift   — audio/mel/encoder-step tensor utilities + findMaxIndex

tokenizerPiece bumped private -> internal (now called cross-file).

Verified byte-identical: 100-sample test-clean A/B unchanged (WER 4.0% /
CER 1.3%, 0/100 transcripts changed). Builds clean arm64 + x86_64.
